PostIssueingHooks -> PostIssuingHooks
Renaming modules with deprecation, and backward compatibility for
transition.
