+++ lib/acmesmith/post_issuing_hooks/base.rb
@@ -1,0 +1,21 @@
+module Acmesmith
+  module PostIssuingHooks
+    class Base
+      attr_reader :certificate
+
+      def common_name
+        certificate.common_name
+      end
+
+      def run(certificate:)
+        @certificate = certificate
+        execute
+      end
+
+      def execute
+        raise NotImplementedError
+      end
+    end
+  end
+end
+
